Fix texture size reporting (#17)
* Getting closer to accurate texture sizes

For versions that have the function we need this gets us the size of the texture with the caveat that if it's an auto format you're getting the size of the current platform and not the specified build platform.  Hopefully that is rectified in the next commit.

* Get real format when texture is auto

TextureFormat and TextureImporterFormat match except the latter has a few leading negative entries to handle the auto formats.  In those cases we need to ask for the remap or else results are wrong.

* Update changelog

* Update changelog notes

Based on Macro's feedback.