Anim module (#56)
* Initial files

Not hooked up to the UI and no reporting or analysis yet, but the skeleton is there.

* Controller reporting

This should report all raw property data from the controller asset, will need to hook it up to the UI next to be sure.

* Actually call the reporting functions

Oops!

* Hook up module to UI

* Don't need this TODO anymore

* Add asset module layout

All other asset modules seem to do it, so I guess I should, too.

* Animation clip work

Setting up the skeleton of the clip reporting code.  Need to fill out the layout to then have this create issues.  Also added logging to controllers when the asset doesn't resolve properly.

* Animation clip reporting

That should be everything.  Need to bring in some content and validate but then the core module should be complete.

It would be nice to do something with bounds but that will require more complex type handling.

* Avatar mask reporting

* Better progress handling

* Small formatting fix

* More formatting fix

* Add to clip reporting

Check if it loops from the base class.

* Add to controller reporting

Get number of clips from base class.

* Avatar reporting

* Add to avatar reporting

Grab everything from the human description.

* Fix version errors

Only check the avatar human description stats on 2019.1 and above.

* Really fix that error

Need to also conditionally define the enum itself or else the issue creation has a mismatch.

* Update the changelog