First pass at adding arbitrary DiagnosticParams to ProjectAuditorRules
From proposal: The proposed name for the second class is ProjectAuditorRules . This class encompasses the functionality of the other 3 classes described above. It contains the mute rules from ProjectAuditorConfig . It contains functionality similar to ProjectAuditorDiagnosticParamsProvider , with a default set of DiagnosticParams and a number of optional overrides. These DiagnosticParams are not quite the same as the old ProjectAuditorDiagnosticParams, and are instead simply a serialized list of <string, int> key/value pairs, which can be used to build dictionaries after the object has been loaded and deserialized. The default params in a newly-created ProjectAuditorRules are initialized with names and values that match those that are currently in ProjectAuditorDiagnosticParams. ProjectAuditorRules will provide methods for getting the value for a given key (checking for overrides from the selected DiagnosticParams first before falling back on the default) which can be used by the analyzers.