ProjectAuditorParams gets constructed with (or finds, or creates) ProjectAuditorRules.
ProjectAuditorParams will contain a reference to ProjectAuditorRules (as a replacement for the ProjectAuditorDiagnosticParams it contains currently), which in turn will provide values from the selected DiagnosticParams. ProjectAuditor will no longer have the option to be initialized with a path to any config file - all the information it requires will be passed in via ProjectAuditorParams. ProjectAuditorDiagnosticParams and ProjectAuditorDiagnosticParamsProvider are removed.